[project-name]/requirements.txt
Flask
PyGithub
requests